reorganize existing subpackages and modules
