+++ src/almanack/checks/index.py
@@ -1,0 +1,5 @@
+"""
+Creates and manages an index for the Almanack checks.
+"""
+
+pass
