+++ src/almanack/git.py
